chore(agents): move hooks from components/Agents/ to hooks/
Per the repo's organization (CLAUDE.md: "hooks/ - React hooks for
data fetching and state"), hooks live at top-level. The agent
templates surface was inconsistent — useAgentData, useAgentCard,
useAgentToggleFavorite, and the three new mutation hooks were
co-located in components/Agents/.

Moves all six via git mv (history preserved):
- hooks/useAgentData.ts
- hooks/useAgentCard.ts
- hooks/useAgentToggleFavorite.ts
- hooks/useCreateAgentTemplate.ts
- hooks/useUpdateAgentTemplate.ts
- hooks/useDeleteAgentTemplate.ts

Imports rewritten across 8 call sites (Agents.tsx, TagSelector.tsx,
AgentCard.tsx, CreateAgentDialog.tsx, AgentEditDialog.tsx,
AgentDeleteButton.tsx, components/Chat/StarterAgentCard.tsx,
components/Sidebar/AgentsNavItem.tsx) to use @/hooks/* alias.
