+++ components/Agents/Agents.tsx
@@ -2,9 +2,9 @@ import { useRouter } from "next/navigation";
 import React from "react";
 import AgentTags from "./AgentTags";
 import AgentCard from "./AgentCard";
-import { useAgentData } from "./useAgentData";
-import { useAgentToggleFavorite } from "./useAgentToggleFavorite";
-import type { Agent } from "./useAgentData";
+import { useAgentData } from "@/hooks/useAgentData";
+import { useAgentToggleFavorite } from "@/hooks/useAgentToggleFavorite";
+import type { Agent } from "@/hooks/useAgentData";
 import CreateAgentButton from "./CreateAgentButton";
 import { Switch } from "@/components/ui/switch";
 import AgentsSkeleton from "./AgentsSkeleton";
