+++ hooks/useAgentToggleFavorite.ts
