+++ hooks/useDeleteAgentTemplate.ts
